Add files via upload
I had written a code for the stocks which are moving in range bound, based on ATR(7) values. I considered daily time frame and 15 days of data. The ATR value should be in between 5 and -5. The logic is, it will compare the ATR of the last candle to the previous candlesticks, upto 10 candles. If the ATR is in-between these values, it triggers that stock is in consolidation phase or else, the stock is following a trend.