Add files via upload
Finding out the Correlation values based on daily closing values. Multiplying the Corr values by 100. Then adding all the correlation (ABSOLUTE) values. It has 543 stocks. If the summation is zero, it means it most UNCORRELATED. If the  sum of Absolute values is equal to 543, it means, it is Highly CORRELATED. The minimum values I got, is  around 88. These are uncorrelated stocks.  Then, the maximum values are around 330. Which means, these are highly CORRELATED.