Add files via upload
I had considered daily volumes along with OPHC values. I had found the difference of High and low, as well as Close and open. I had divided the volume by these diffrences and plotted the. Some of the plots shows abnormal patterns. We can then focus on these plots. We need to sort them in ascending order. The first and last values , and the dates of these values are to be considered for entries. Risk 1%, reward 1.5%, and duration of the trades can be in a range of 2-3 days.